move DateLabel to own component
