+++ src/features/calendar/components/CalendarDayView/Day/index.tsx
@@ -1,13 +1,10 @@
-import dayjs from 'dayjs';
-import { FormattedDate } from 'react-intl';
-import { Box, Typography } from '@mui/material';
+import { Box } from '@mui/material';
 
+import DateLabel from './DateLabel';
 import { DayInfo } from '../types';
 import Event from './Event';
-import theme from 'theme';
 
 const Day = ({ date, dayInfo }: { date: Date; dayInfo: DayInfo }) => {
-  const isToday = dayjs(date).isSame(dayjs(), 'day');
   return (
     <Box
       alignItems="flex-start"
@@ -19,27 +16,7 @@ const Day = ({ date, dayInfo }: { date: Date; dayInfo: DayInfo }) => {
         backgroundColor: '#eeeeee',
       }}
     >
-      {/* Date */}
-      <Box
-        alignItems="center"
-        display="flex"
-        gap={1}
-        sx={{
-          backgroundColor: isToday ? theme.palette.primary.main : undefined,
-          borderRadius: '20%/50%',
-          color: isToday ? 'white' : 'inherit',
-          padding: '10px 15px',
-        }}
-      >
-        <Typography fontSize="1.4em" variant="h4">
-          <FormattedDate day="numeric" value={date} />
-        </Typography>
-        <Typography variant="body1">
-          <FormattedDate month="long" value={date} />
-          ,&nbsp;
-          <FormattedDate value={date} weekday="short" />
-        </Typography>
-      </Box>
+      <DateLabel date={date} />
       {/* Remaining space for list of events */}
       <Box
         display="flex"
